Drive the placement guard tests through the public Map API
